+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 Unreleased
 ************
+* [BUGFIX] Pass scan validation if no SPWs are defined on a CSP configuration
 
 12.0.0
 **********